+++ resources/scripts/install-bun.js
@@ -2,8 +2,8 @@ const fs = require('fs')
 const path = require('path')
 const os = require('os')
 const { execSync } = require('child_process')
-const https = require('https')
 const AdmZip = require('adm-zip')
+const { downloadWithRedirects } = require('./download')
 
 // Base URL for downloading bun binaries
 const BUN_RELEASE_BASE_URL = 'https://github.com/oven-sh/bun/releases/download'
@@ -22,41 +22,6 @@ const BUN_PACKAGES = {
   'linux-musl-x64': 'bun-linux-x64-musl.zip',
   'linux-musl-x64-baseline': 'bun-linux-x64-musl-baseline.zip',
   'linux-musl-arm64': 'bun-linux-aarch64-musl.zip'
-}
-
-/**
- * Downloads a file from a URL with redirect handling
- * @param {string} url The URL to download from
- * @param {string} destinationPath The path to save the file to
- * @returns {Promise<void>}
- */
-async function downloadWithRedirects(url, destinationPath) {
-  return new Promise((resolve, reject) => {
-    const file = fs.createWriteStream(destinationPath)
-    const request = (url) => {
-      https
-        .get(url, (response) => {
-          if (response.statusCode === 302 || response.statusCode === 301) {
-            // Handle redirect
-            request(response.headers.location)
-            return
-          }
-
-          if (response.statusCode !== 200) {
-            reject(new Error(`Failed to download: ${response.statusCode}`))
-            return
-          }
-
-          response.pipe(file)
-          file.on('finish', () => {
-            file.close(resolve)
-          })
-        })
-        .on('error', reject)
-    }
-
-    request(url)
-  })
 }
 
 /**
